Split SKFourByteTag tests into dedicated test file
SKFourByteTagTest.cs: 14 tests for Parse, ToString, conversions,
equality, edge cases. These test a standalone type, not SKTypeface.
